refactor(exif): ExifService の読み取り責務を ExifReader へ分離 (#199)
- GetMetadataAsync / GetMetadata / ReadMetadata を ExifReader（internal static）へ移動し、MetadataExtractor 依存を局所化
- ExifService は書き込み専用に縮退（558 行 → 380 行）
- 呼び出し側 5 箇所を ExifReader.GetMetadataAsync へ振り向け
- 読み取り系テストを ExifReaderTests へ再編し、非画像 / 非存在ファイルの null 返却ケースを追加
