+++ PhotoGeoExplorer.Core/Services/ExifService.cs
@@ -1,35 +1,20 @@
 using System;
-using System.Collections.Generic;
 using System.Globalization;
 using System.IO;
-using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
-using MetadataExtractor;
-using MetadataExtractor.Formats.Exif;
-using MetadataExtractor.Formats.FileSystem;
-using PhotoGeoExplorer.Models;
 using SixLabors.ImageSharp;
 using SixLabors.ImageSharp.Metadata.Profiles.Exif;
 using ImageSharpRational = SixLabors.ImageSharp.Rational;
 
 namespace PhotoGeoExplorer.Services;
 
+/// <summary>
+/// EXIF メタデータの書き込み責務を担う（読み取りは <see cref="ExifReader"/> を参照）
+/// </summary>
 internal static class ExifService
 {
     private static readonly byte[] ExifHeader = { 0x45, 0x78, 0x69, 0x66, 0x00, 0x00 };
-
-    public static Task<PhotoMetadata?> GetMetadataAsync(string filePath, CancellationToken cancellationToken)
-    {
-        ArgumentNullException.ThrowIfNull(filePath);
-        return Task.Run(() => ReadMetadata(filePath, cancellationToken), cancellationToken);
-    }
-
-    public static PhotoMetadata? GetMetadata(string filePath)
-    {
-        ArgumentNullException.ThrowIfNull(filePath);
-        return ReadMetadata(filePath, CancellationToken.None);
-    }
 
     public static Task<bool> UpdateMetadataAsync(
         string filePath,
@@ -41,103 +26,6 @@ internal static class ExifService
     {
         ArgumentNullException.ThrowIfNull(filePath);
         return Task.Run(() => WriteMetadata(filePath, takenAt, latitude, longitude, updateFileModifiedDate, cancellationToken), cancellationToken);
-    }
-
-    private static PhotoMetadata? ReadMetadata(string filePath, CancellationToken cancellationToken)
-    {
-        cancellationToken.ThrowIfCancellationRequested();
-
-        IReadOnlyList<MetadataExtractor.Directory> directories;
-        try
-        {
-            directories = ImageMetadataReader.ReadMetadata(filePath);
-        }
-        catch (MetadataExtractor.ImageProcessingException ex)
-        {
-            AppLog.Error($"Failed to read metadata: {filePath}", ex);
-            return null;
-        }
-        catch (IOException ex)
-        {
-            AppLog.Error($"Failed to read metadata: {filePath}", ex);
-            return null;
-        }
-        catch (UnauthorizedAccessException ex)
-        {
-            AppLog.Error($"Failed to read metadata: {filePath}", ex);
-            return null;
-        }
-        catch (NotSupportedException ex)
-        {
-            AppLog.Error($"Failed to read metadata: {filePath}", ex);
-            return null;
-        }
-#pragma warning disable CA1031 // Do not catch general exception types
-        catch (Exception ex)
-        {
-            // Catch any other unexpected exceptions from MetadataExtractor library
-            // (e.g., IndexOutOfRangeException when processing certain MP3 files)
-            // to prevent app crashes. These are logged and treated as metadata read failures.
-            AppLog.Error($"Unexpected exception reading metadata: {filePath}", ex);
-            return null;
-        }
-#pragma warning restore CA1031 // Do not catch general exception types
-
-        cancellationToken.ThrowIfCancellationRequested();
-
-        try
-        {
-            var gpsDirectory = directories.OfType<GpsDirectory>().FirstOrDefault();
-            GeoLocation? location = null;
-            if (gpsDirectory is not null && gpsDirectory.TryGetGeoLocation(out var geoLocation))
-            {
-                location = geoLocation;
-            }
-
-            double? latitude = location?.Latitude;
-            double? longitude = location?.Longitude;
-
-            var exifIfd0 = directories.OfType<ExifIfd0Directory>().FirstOrDefault();
-            var cameraMake = exifIfd0?.GetString(ExifDirectoryBase.TagMake);
-            var cameraModel = exifIfd0?.GetString(ExifDirectoryBase.TagModel);
-
-            var subIfd = directories.OfType<ExifSubIfdDirectory>().FirstOrDefault();
-            DateTime? dateTime = null;
-            if (subIfd is not null)
-            {
-                if (subIfd.TryGetDateTime(ExifDirectoryBase.TagDateTimeOriginal, out var dt))
-                    dateTime = dt;
-                else if (subIfd.TryGetDateTime(ExifDirectoryBase.TagDateTimeDigitized, out dt))
-                    dateTime = dt;
-            }
-
-            if (dateTime is null)
-            {
-                var fileMeta = directories.OfType<FileMetadataDirectory>().FirstOrDefault();
-                if (fileMeta is not null && fileMeta.TryGetDateTime(FileMetadataDirectory.TagFileModifiedDate, out var dt))
-                    dateTime = dt;
-            }
-
-            DateTimeOffset? takenAt = null;
-            if (dateTime.HasValue)
-            {
-                var localTime = DateTime.SpecifyKind(dateTime.Value, DateTimeKind.Local);
-                takenAt = new DateTimeOffset(localTime);
-            }
-
-            return new PhotoMetadata(
-                takenAt,
-                cameraMake,
-                cameraModel,
-                latitude,
-                longitude,
-                hasGpsData: gpsDirectory is not null);
-        }
-        catch (MetadataExtractor.MetadataException ex)
-        {
-            AppLog.Error($"Partial metadata read failure: {filePath}", ex);
-            return null;
-        }
     }
 
     private static bool WriteMetadata(
